Fix retrieving history when the entity was unavailable
If the entity was unavailable at some point, it seems we get
`unavailable` returned in its history. Currently, we'll try to progress
this value as if it were a number, and we get:

> Could not convert 'unavailable' to float in check_ambient_air_temperature()

in the logs. We can suppress this by filtering out such values before
trying to process them.